Добавил конструкцию try catch на удаление файлов

+++ Module_8_FinalExercises/Task_3/Program.cs
@@ -10,13 +10,13 @@
         //Показать, сколько папка весит после очистки.
         static void Main(string[] args)
         {
-            string dirPath = @"D:\Programming\Skillfactory\C#_projects\Module_8_FinalExercises\Module_8_FinalExercises\Task_3\FolderForTask3\TestFolder";
+            string dirPath = @"D:\Programming\Skillfactory\C#_projects\Module_8\Module_8_FinalExercises\Module_8_FinalExercises\Task_3\FolderForTask3\TestFolder";
             int fileCount = 0;
             DirectoryInfo directory = new DirectoryInfo(dirPath);
 
-            DirectorySize(directory,"Исходный размер папки: ",ref fileCount);
-            int deletedFileCount=fileCount;
-            
+            DirectorySize(directory, "Исходный размер папки: ", ref fileCount);
+            int deletedFileCount = fileCount;
+
             if (directory.Exists)
             {
                 long deletedSize = DirectoryAndFileSize(directory, ref fileCount);
@@ -24,31 +24,58 @@
                 Console.WriteLine($"Освобождено: {deletedSize} байт. Удалено: {deletedFileCount} файлов");
                 DirectorySize(directory, "Текущий размер папки: ", ref fileCount);
             }
-            else 
+            else
             {
                 Console.WriteLine("Удаление невозможно");
             }
 
-            
+
         }
         static void DeleteInFolder(DirectoryInfo directory)
         {
+            foreach (FileInfo file in directory.GetFiles())
+            {
+                try
+                {
+                    if (file.Exists) file.Delete();
+                    else throw new FileNotFoundException("Файла не существует");
+                }
+                catch (FileNotFoundException ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
 
-                    foreach (FileInfo file in directory.GetFiles())
-                    {
-                        file.Delete();
-                    }          
-                    foreach (DirectoryInfo dir in directory.GetDirectories())
-                    {
-                        dir.Delete(true);
-                    }                   
+
+            }
+            foreach (DirectoryInfo dir in directory.GetDirectories())
+            {
+                try
+                {
+                    if (dir.Exists) dir.Delete(true);
+                    else throw new DirectoryNotFoundException("Папки не существует");
+                }
+                catch (DirectoryNotFoundException ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+
+            }
         }
-        static void DirectorySize(DirectoryInfo directory, string message,ref int fileCount)
+
+        static void DirectorySize(DirectoryInfo directory, string message, ref int fileCount)
         {
-            
+
             if (directory.Exists)
             {
-                Console.WriteLine($"{message}{DirectoryAndFileSize(directory,ref fileCount)} байт");
+                Console.WriteLine($"{message}{DirectoryAndFileSize(directory, ref fileCount)} байт");
             }
             else
             {
@@ -56,7 +83,7 @@
             }
         }
 
-        static long DirectoryAndFileSize(DirectoryInfo directory,ref int fileCount)
+        static long DirectoryAndFileSize(DirectoryInfo directory, ref int fileCount)
         {
             long dirSize = 0;
             foreach (FileInfo file in directory.GetFiles())
